pyproject.toml
[project]
name = "exan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaleido>=1.0.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-box>=7.3.2",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]
